requirements.txt
apiclient==1.0.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
DAWG-Python==0.7.2
docopt==0.6.2
google-api-core==2.1.0
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.2
iniconfig==1.1.1
intervaltree==3.1.0
ipymarkup==0.9.0
joblib==1.1.0
lxml==4.6.3
natasha==1.4.0
navec==0.10.0
nltk==3.6.3
numpy==1.21.2
oauth2client==4.1.3
packaging==21.0
pluggy==1.0.0
protobuf==3.18.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pytest==6.2.5
razdel==0.5.0
regex==2021.10.8
requests==2.26.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
selenium==3.141.0
six==1.16.0
sklearn==0.0
slovnet==0.5.0
sortedcontainers==2.4.0
soupsieve==2.2.1
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
uritemplate==4.1.0
urllib3==1.26.7
webdriver-manager==3.5.0
yargy==0.15.0
requests==2.24.0
mock==4.0.3
pymorphy2==0.9.1
aiohttp==3.8.1
aiosignal==1.2.0
apiclient==1.0.4
async-timeout==4.0.1
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.11.0
DAWG-Python==0.7.2
docopt==0.6.2
fonttools==4.28.1
frozenlist==1.2.0
future==0.18.2
google-api-core==2.1.0
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
graphviz==0.5.1
httplib2==0.20.1
idna==3.2
iniconfig==1.1.1
intervaltree==3.1.0
ipymarkup==0.9.0
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.5.0
multidict==5.2.0
multiprocessing-logging==0.3.1
natasha==1.4.0
navec==0.10.0
neupy==0.6.5
nltk==3.6.3
numpy==1.21.2
oauth2client==4.1.3
packaging==21.0
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
progressbar2==3.34.3
protobuf==3.18.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
razdel==0.5.0
regex==2021.10.8
requests==2.26.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
slovnet==0.5.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tableprint==0.7.1
Theano==1.0.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.1
uritemplate==4.1.0
urllib3==1.26.7
wcwidth==0.2.5
yargy==0.15.0
yarl==1.7.2
